Fixing Quality Gate fail (#7)

+++ src/main/java/org/apache/commons/text/StrBuilder.java
@@ -151,7 +151,7 @@ public class StrBuilder implements CharSequence, Appendable, Serializable, Build
         @Override
         public long skip(long n) {
             if (pos + n > size()) {
-                n = size() - ((long)pos);
+                n = size() - pos;
             }
             if (n < 0) {
                 return 0;
